Move version related functions to peers submodule
